+++ database/src/lib.rs
@@ -335,6 +335,7 @@ fn column_mapping(
         | Type::Option(_)
         | Type::Result(_, _)
         | Type::Array(_)
+        | Type::HttpResult(_)
         | Type::Unit
         | Type::Unknown => "TEXT",
     };